version 1.7.05
Made a Chrome packaged app, which required a lot of changes since they
don’t allow in-line javascript. The code here is essentially the same,
except that the javascript code is added back at the head and the end
of the body rather than kept in separate .js files (needed for SHA256
code verification).

Features added:

1. Bigger checkboxes and radio buttons
2. For the Chrome extension: Lock database syncs between devices using
chrome.storage.